pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool.poetry]
name = "bluelamp-ai"
version = "1.3.1"
description = "BlueLamp AI: 日本語対応AIエージェントシステム - オーケストレーターと拡張マネージャー（最新版）"
readme = "README_PUBLIC.md"
authors = [ "BlueLamp Team" ]
license = "MIT"
repository = "https://github.com/your-repo/bluelamp-ai"
keywords = ["ai", "agent", "japanese", "automation", "development"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Natural Language :: Japanese",
]
packages = [
  { include = "core" },
  { include = "extensions" },
  { include = "resources" },
  { include = "pyproject.toml" },
  { include = "poetry.lock" },
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
litellm = "^1.60.0, !=1.64.4, !=1.67.*"            # avoid 1.64.4 (known bug) & 1.67.* (known bug #10272)
aiohttp = ">=3.9.0,!=3.11.13"                      # Pin to avoid yanked version 3.11.13
google-generativeai = "*"                          # To use litellm with Gemini Pro API
google-api-python-client = "^2.164.0"              # For Google Sheets API
google-auth-httplib2 = "*"                         # For Google Sheets authentication
google-auth-oauthlib = "*"                         # For Google Sheets OAuth
termcolor = "*"
toml = "*"
types-toml = "*"
numpy = "*"
json-repair = "*"

html2text = "*"
pexpect = "*"
jinja2 = "^3.1.3"
fastapi = "*"
docker = "*"
tenacity = ">=8.5,<10.0"
zope-interface = "7.2"
pathspec = "^0.12.1"
pyjwt = "^2.9.0"
dirhash = "*"
python-dotenv = "*"
rapidfuzz = "^3.9.0"
whatthepatch = "^1.0.6"
protobuf = "^5.0.0,<6.0.0"                         # Updated to support newer opentelemetry
opentelemetry-api = "^1.33.1"
opentelemetry-exporter-otlp-proto-grpc = "^1.33.1"
libtmux = ">=0.37,<0.40"
pygithub = "^2.5.0"
joblib = "*"
sse-starlette = "^2.1.3"
psutil = "*"
python-json-logger = "^3.2.1"
prompt-toolkit = "^3.0.50"
# poetry = "^2.1.2"  # 削除: 循環依存回避
anyio = "4.9.0"
pythonnet = "*"
fastmcp = "^2.5.2"
python-frontmatter = "^1.1.0"
# Runtime dependencies moved to separate group
bashlex = "^0.18"
openhands_aci = "0.3.0"

# Cloud integrations moved to separate group
pyyaml = "^6.0.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "0.12.0"
mypy = "1.16.1"
pre-commit = "4.2.0"
build = "*"
types-setuptools = "*"
pytest = "^8.4.0"

# Document processing dependencies (optional)
[tool.poetry.group.runtime]
optional = true

[tool.poetry.group.runtime.dependencies]
ipywidgets = "^8.1.5"
qtconsole = "^5.6.1"
PyPDF2 = "*"
python-pptx = "*"
pylatexenc = "*"
python-docx = "*"

# Cloud service integrations (optional)
[tool.poetry.group.integrations]
optional = true

[tool.poetry.group.integrations.dependencies]
redis = ">=5.2,<7.0"
minio = "^7.2.8"
daytona = "0.21.1"
stripe = ">=11.5,<13.0"
google-cloud-aiplatform = "*"
anthropic = { extras = [ "vertex" ], version = "*" }
boto3 = "*"
kubernetes = "^33.1.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-cov = "*"
pytest-asyncio = "*"
pytest-forked = "*"
pytest-xdist = "*"
openai = "*"
pandas = "*"
reportlab = "*"
gevent = ">=24.2.1,<26.0.0"



[tool.poetry.scripts]
# 開発用内部コマンド（一般配布では非表示）
openhands = "extensions.cli.main:main"
openhands2 = "extensions.cli.main_delegation:main"
openhands3 = "extensions.cli.main_session:main"
# 一般公開用日本語コマンド
"ブルーランプ" = "extensions.cli.bluelamp_public:main"
"ブルーランプ拡張" = "extensions.cli.bluelamp_extension_public:main"



[tool.poetry-dynamic-versioning]
enable = true
style = "semver"

[tool.autopep8]
# autopep8 fights with mypy on line length issue
ignore = [ "E501" ]

[tool.black]
# prevent black (if installed) from changing single quotes to double quotes
skip-string-normalization = true

[tool.ruff]
lint.select = [ "D" ]
# ignore warnings for missing docstrings and CLI-irrelevant docstring formatting
lint.ignore = [ "D1", "D205", "D417" ]
lint.pydocstyle.convention = "google"

[tool.coverage.run]
concurrency = [ "gevent" ]
